requirements-prod.txt
# Production dependencies only
fastapi==0.115.6
motor==3.6.0
pandas==2.2.3
pydantic==2.10.3
pydantic-settings==2.6.1
pymongo==4.9.2
requests==2.32.4
prometheus_client~=0.21.1
pymorphy3~=2.0.3
colorama~=0.4.6
uvicorn~=0.32.1
numpy~=2.1.3
starlette~=0.41.0
psutil~=5.9.0
python-dotenv~=1.0.0
openpyxl~=3.1.0
